requirements.txt
anyio==3.6.1
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
bcrypt==3.2.2
botocore==1.27.27
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
executing==0.8.3
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
idna==3.3
ipdb==0.13.9
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
orjson==3.7.7
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
PyJWT==2.4.0
pymongo==4.1.1
pymongo-auth-aws==1.0.2
pymongocrypt==1.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
stack-data==0.3.0
starlette==0.19.1
toml==0.10.2
traitlets==5.3.0
typing-extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
